Split documents features of impress_maintenance into maintenance_documents
