+++ impress_maintenance/__manifest__.py
@@ -26,7 +26,6 @@
         "views/quality/maintenance_equipment_quality_action.xml",
         "views/menus.xml",
         "security/maintenance.xml",
-        "data/document_folder.xml",
     ],
     "application": True,
     "installable": True,
