+++ impress_maintenance/models/__init__.py
@@ -3,4 +3,3 @@ from . import maintenance_request
 from . import maintenance_team
 from . import maintenance_equipment_category
 from . import maintenance_equipment
-from . import documents_document
