+++ maintenance_documents/__manifest__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+{
+    "name": "Maintenance_documents",
+    "version": "17.0.0.0.1",
+    "summary": """ Bridge module between Maintenance and Documents """,
+    "author": "Cédric Paradis",
+    "website": "",
+    "category": "Hidden",
+    "depends": [
+        "base",
+        "documents",
+        "maintenance_worksheet",
+    ],
+    "data": [
+        "views/maintenance_equipment_views.xml"
+    ],
+    "installable": True,
+    "auto_install": False,
+    "license": "GPL-2",
+}
