+++ maintenance_documents/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import documents_document
+from . import maintenance_equipment
